fix(ascend): fix sum global reduce, add abs/acos ops, fix nll_loss output shape
- sum.cc: fix global reduce (no dim or empty dim) passing empty array to ACL
  causing ret=161002; now fills all dims automatically
- abs.cc: add abs op Ascend backend implementation
- acos.cc: add acos op Ascend backend with float64->float32 downcast
- op_api_common.h: add kDouble -> ACL_DOUBLE mapping in ToAclDataType
- nll_loss.cc: fix output/total_weight scalar shape causing ACL ret=161001;
  use shape{1} then squeeze back to scalar
- backends_ascend.conf: register abs, acos to ascend backend
- test_acos_dispatch.py: change float64 param to float16 (device rand fp64 bug)

+++ tests/integration/ops/test_acos_dispatch.py
@@ -74,7 +74,7 @@ class TestAcosCorrectness:
         ref = torch.acos(a.cpu())
         torch.testing.assert_close(out.cpu(), ref, rtol=1e-5, atol=1e-5)
 
-    @pytest.mark.parametrize("dtype", [torch.float32, torch.float64])
+    @pytest.mark.parametrize("dtype", [torch.float32, torch.float16])
     @pytest.mark.anyplatform
     def test_acos_dtypes(self, dtype):
         torch.manual_seed(2)
